feat : Main Page  Card 구현 Main Page 중단 영화 카드 레이아웃 구현
