+++ client/src/routes/Home.js
@@ -10,28 +10,38 @@ const Wrapper = styled.div`
   margin: 10rem auto;
   justify-content: center;
   align-items: center;
+  flex-direction: column;
 `;
 
 const Title = styled.h1`
   font-size: 2rem;
   font-weight: 600;
   text-align: center;
-  margin-right: 2rem;
+  margin-right: 1rem;
 `;
 
+const Header = styled.div`
+  display:flex;
+  width:100%;
+  justify-content : center;
+  align-items:center;
+`
 
 
 export default function Home() {
   return (
     <>
       <Layout isNavSearch={true} isMain={true} />
-
       <Wrapper>
+        <Header>
         <Title>Title</Title>
         <SearchInput/>
+        </Header>
+        <HomeCard/>
+        <HomeCard/>
       </Wrapper>
 
-      <HomeCard/>
+      
     </>
   );
 }
